+++ superduper/components/template.py
@@ -86,7 +86,9 @@ class Template(Component):
     def postinit(self):
         """Post initialization method."""
         if isinstance(self.template, Base):
-            self.template = self.template.encode(defaults=True, metadata=False)
+            self.template = self.template.encode(
+                defaults=True, metadata=False, cfg=CFG(json_native=True)
+            )
 
         if '_blobs' in self.template:
             self.blobs = self.template.pop('_blobs')