+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Optimize performance during apply.
 - Add option of multiple `VectorIndex` implementations
 - Batch create tables, put-components
+- Add fix for exporting component with `CFG.json_native=True`
 
 #### New Features & Functionality